use file util and add local files option for import
