+++ etl/cad_incidents_import/incidents_to_s3.py
@@ -4,9 +4,10 @@ Transfer CAD data CSV files from COACD network location to AWS S3 inbox
 
 import argparse
 import os
-import re
 import boto3
 import logging
+
+from utils.files import get_local_files_to_process
 
 logging.basicConfig(
     level=logging.INFO,
@@ -14,38 +15,10 @@ logging.basicConfig(
 )
 logger = logging.getLogger(__name__)
 
-COACD_MOUNT_PATH = os.environ.get("COACD_MOUNT_PATH", "/mnt/tpw_cadtraffic")
+COACD_MOUNT_PATH = os.environ.get("COACD_MOUNT_PATH", "/mnt/vision_zero_cad")
 BUCKET_NAME = os.environ["BUCKET_NAME"]
 BUCKET_ENV = os.environ["BUCKET_ENV"]
 S3_PREFIX = f"{BUCKET_ENV}/cad_incidents/inbox"
-
-
-def is_file_to_process(filename):
-    """
-    Check if the file name matches our naming convention
-    - TPWCADTrafficSafetyWithGroupIDDaily_20260410.CSV
-    - TPWCADTrafficSafetyDaily_20260410.CSV
-    """
-    return (
-        filename.lower().endswith(".csv")
-        and (
-            "TPWCADTrafficSafetyWithGroupIDDaily" in filename
-            or "TPWCADTrafficSafetyDaily" in filename
-        )
-        and len(filename.split("_")) == 2
-    )
-
-
-def extract_date_from_filename(filename: str) -> str:
-    """Extract the date string from a filename for sorting.
-
-    We expect files named in either of these two ways
-        - TPWCADTrafficSafetyWithGroupIDDaily_20260410.CSV
-        - TPWCADTrafficSafetyDaily_20260410.CSV
-    """
-    full_name, ext = filename.split(".")
-    name, dt = full_name.split("_")
-    return dt
 
 
 def main():
@@ -58,7 +31,7 @@ def main():
         help="Log what would be uploaded and deleted without actually doing it",
     )
     parser.add_argument(
-        "--skip-archive",
+        "--skip-removal",
         action="store_true",
         help="Don't delete the file from the network drive after processing",
     )
@@ -69,35 +42,32 @@ def main():
 
     s3 = boto3.client("s3")
 
-    files = [f for f in os.listdir(COACD_MOUNT_PATH) if is_file_to_process(f)]
+    files_todo = get_local_files_to_process(dir_name=COACD_MOUNT_PATH)
 
-    if not files:
+    if not files_todo:
         raise Exception("No files found in COACD network directory")
 
-    files.sort(key=extract_date_from_filename)
-    logger.info(f"Found {len(files)} file(s) to process.")
+    logger.info(f"Found {len(files_todo)} file(s) to process.")
 
-    for filename in files:
-        local_path = os.path.join(COACD_MOUNT_PATH, filename)
+    for filename in files_todo:
+        file_path = os.path.join(COACD_MOUNT_PATH, filename)
         s3_key = f"{S3_PREFIX}/{filename}"
 
         if args.dry_run:
-            logger.info(
-                f"[DRY RUN] Would upload {filename} → s3://{BUCKET_NAME}/{s3_key}"
-            )
+            logger.info(f"[DRY RUN] Would upload s3://{BUCKET_NAME}/{s3_key}")
 
         else:
-            logger.info(f"Uploading {filename} → s3://{BUCKET_NAME}/{s3_key}")
-            s3.upload_file(local_path, BUCKET_NAME, s3_key)
+            logger.info(f"Uploading s3://{BUCKET_NAME}/{s3_key}")
+            s3.upload_file(file_path, BUCKET_NAME, s3_key)
 
-        if not args.skip_archive:
+        if not args.skip_removal:
             if args.dry_run:
-                logger.info(f"[DRY RUN] Would remove {local_path}")
+                logger.info(f"[DRY RUN] Would remove {file_path}")
             else:
-                logger.info(f"Removing {local_path}")
+                logger.info(f"Removing {file_path}")
                 # os.remove(local_path)
 
-    logger.info(f"Done. Processed {len(files)} file(s).")
+    logger.info(f"Done. Processed {len(files_todo)} file(s).")
 
 
 if __name__ == "__main__":
